fix(stache): run targeted item updates on BasicStore
cacheItem, forgetItem, and getKeyFromPath already live there. The new methods belong next to them, not on Store.
