+++ .phpstan/baseline.neon
@@ -195,19 +195,19 @@ parameters:
 		-
 			message: '#^Call to an undefined method Statamic\\Stache\\Stores\\Store\:\:cacheItem\(\)\.$#'
 			identifier: method.notFound
-			count: 3
+			count: 2
 			path: ../src/Stache/Stores/Store.php
 
 		-
 			message: '#^Call to an undefined method Statamic\\Stache\\Stores\\Store\:\:forgetItem\(\)\.$#'
 			identifier: method.notFound
-			count: 4
+			count: 2
 			path: ../src/Stache/Stores/Store.php
 
 		-
 			message: '#^Call to an undefined method Statamic\\Stache\\Stores\\Store\:\:getKeyFromPath\(\)\.$#'
 			identifier: method.notFound
-			count: 2
+			count: 1
 			path: ../src/Stache/Stores/Store.php
 
 		-
