+++ src/Stache/Stores/BasicStore.php
@@ -2,8 +2,11 @@
 
 namespace Statamic\Stache\Stores;
 
+use Illuminate\Support\Enumerable;
 use Statamic\Facades\File;
+use Statamic\Facades\Path;
 use Statamic\Facades\Stache;
+use Statamic\Support\Arr;
 use Symfony\Component\Finder\SplFileInfo;
 
 abstract class BasicStore extends Store
@@ -111,6 +114,69 @@ abstract class BasicStore extends Store
     protected function getKeyFromPath($path)
     {
         return $this->paths()->flip()->get($path);
+    }
+
+    public function updateItemFromPath(string $path): void
+    {
+        foreach (Arr::wrap($this->getItemFromModifiedPath($path)) as $item) {
+            $key = $this->getItemKey($item);
+
+            $this->forgetItem($key);
+            $this->setPath($key, $item->path());
+            $this->cacheItem($item);
+            $this->handleModifiedItem($item);
+
+            $this->resolveIndexes()->filter->isCached()->each(function ($index) use ($item) {
+                $index->updateItem($item);
+            });
+        }
+    }
+
+    public function forgetItemByPath(string $path): void
+    {
+        $key = $this->getKeyFromPathVariants($path);
+
+        collect($key)->each(function ($key) use ($path) {
+            $this->forgetItem($key);
+            $this->forgetPath($key);
+            $this->resolveIndexes()->filter->isCached()->each->forgetItem($key);
+            $this->handleDeletedItem($path, $key);
+        });
+    }
+
+    protected function getKeyFromPathVariants(string $path)
+    {
+        foreach ($this->pathLookupVariants($path) as $candidate) {
+            $key = $this->getKeyFromPath($candidate);
+
+            if ($key instanceof Enumerable) {
+                if ($key->isNotEmpty()) {
+                    return $key;
+                }
+
+                continue;
+            }
+
+            if ($key !== null && $key !== false && $key !== '') {
+                return $key;
+            }
+        }
+
+        return null;
+    }
+
+    protected function pathLookupVariants(string $path): array
+    {
+        $tidy = Path::tidy($path);
+        $resolved = Path::resolve($path);
+
+        return array_values(array_unique(array_filter([
+            $path,
+            $tidy,
+            rtrim($tidy, '/'),
+            $resolved,
+            rtrim($resolved, '/'),
+        ])));
     }
 
     public function save($item)
