+++ src/Stache/Stores/Store.php
@@ -3,7 +3,6 @@
 namespace Statamic\Stache\Stores;
 
 use Facades\Statamic\Stache\Traverser;
-use Illuminate\Support\Enumerable;
 use Statamic\Facades\File;
 use Statamic\Facades\Path;
 use Statamic\Facades\Stache;
@@ -289,77 +288,6 @@ abstract class Store
         });
 
         $this->modified = $modified;
-    }
-
-    /**
-     * Update the stache cache for a single file that was added or modified on disk.
-     * This is a read-only stache operation — it does not write to disk.
-     */
-    public function updateItemFromPath(string $path): void
-    {
-        foreach (Arr::wrap($this->getItemFromModifiedPath($path)) as $item) {
-            $key = $this->getItemKey($item);
-
-            $this->forgetItem($key);
-            $this->setPath($key, $item->path());
-            $this->cacheItem($item);
-            $this->handleModifiedItem($item);
-
-            $this->resolveIndexes()->filter->isCached()->each(function ($index) use ($item) {
-                $index->updateItem($item);
-            });
-        }
-    }
-
-    /**
-     * Remove a single item from the stache cache by its file path.
-     * Used when a file has been deleted from disk.
-     */
-    public function forgetItemByPath(string $path): void
-    {
-        $key = $this->getKeyFromPathVariants($path);
-
-        collect($key)->each(function ($key) use ($path) {
-            $this->forgetItem($key);
-            $this->forgetPath($key);
-            $this->resolveIndexes()->filter->isCached()->each->forgetItem($key);
-            $this->handleDeletedItem($path, $key);
-        });
-    }
-
-    protected function getKeyFromPathVariants(string $path)
-    {
-        foreach ($this->pathLookupVariants($path) as $candidate) {
-            $key = $this->getKeyFromPath($candidate);
-
-            if ($key instanceof Enumerable) {
-                if ($key->isNotEmpty()) {
-                    return $key;
-                }
-
-                continue;
-            }
-
-            if ($key !== null && $key !== false && $key !== '') {
-                return $key;
-            }
-        }
-
-        return null;
-    }
-
-    protected function pathLookupVariants(string $path): array
-    {
-        $tidy = Path::tidy($path);
-        $resolved = Path::resolve($path);
-
-        return array_values(array_unique(array_filter([
-            $path,
-            $tidy,
-            rtrim($tidy, '/'),
-            $resolved,
-            rtrim($resolved, '/'),
-        ])));
     }
 
     protected function handleModifiedItem($item)
